Add title to proposals, move view to mod, refactor pcli proposal cmd, fix previous fvk error
