+++ component/src/governance/mod.rs
@@ -3,10 +3,12 @@ mod component;
 pub(crate) mod event;
 mod execute;
 mod metrics;
+mod view;
 
 pub mod proposal;
 pub mod state_key;
 pub mod tally;
 
 pub use self::metrics::register_metrics;
-pub use component::{Governance, View};
+pub use component::Governance;
+pub use view::View;
