+++ pcli/src/command/query/proposal.rs
@@ -20,47 +20,44 @@ use crate::App;
 
 #[derive(Debug, clap::Subcommand)]
 pub enum ProposalCmd {
-    /// List all proposals by number.
-    List {
+    /// List all governance proposals by number.
+    Proposals {
         /// Whether to include proposals which have already finished voting.
         #[clap(short, long)]
         inactive: bool,
     },
+    /// Query for information about a particular proposal.
+    Proposal {
+        /// The proposal id to query.
+        proposal_id: u64,
+        /// The query to ask of it.
+        #[clap(subcommand)]
+        query: PerProposalCmd,
+    },
+}
+
+#[derive(Debug, clap::Subcommand)]
+pub enum PerProposalCmd {
     /// Fetch the details of a proposal, as submitted to the chain.
-    Fetch {
-        /// The number of the proposal to show.
-        proposal_id: u64,
-    },
+    Definition,
     /// Display the current state of a proposal.
-    State {
-        /// The number of the proposal to show.
-        proposal_id: u64,
-    },
+    State,
     /// Display the voting period of a proposal.
-    Period {
-        /// The number of the proposal to show.
-        proposal_id: u64,
-    },
+    Period,
     /// Display the latest epoch's tally of votes on the proposal, in units of voting power.
-    Tally {
-        /// The number of the proposal to show.
-        proposal_id: u64,
-    },
+    Tally,
     /// List the votes of the validators who have voted on a proposal.
-    ValidatorVotes {
-        /// The number of the proposal to display validator votes for.
-        proposal_id: u64,
-    },
+    ValidatorVotes,
 }
 
 impl ProposalCmd {
     pub async fn exec(&self, app: &mut App) -> Result<()> {
-        use ProposalCmd::*;
+        use PerProposalCmd::*;
 
         let mut client = app.specific_client().await?;
 
         match self {
-            List { inactive } => {
+            ProposalCmd::Proposals { inactive } => {
                 let list: Vec<u64> = if *inactive {
                     let latest: u64 = client.key_proto(latest_proposal_id()).await?;
                     (0..=latest).collect()
@@ -71,49 +68,53 @@ impl ProposalCmd {
                 };
                 json(&list)?;
             }
-            Fetch { proposal_id } => {
-                let description: String =
-                    client.key_proto(proposal_description(*proposal_id)).await?;
-                let payload: ProposalPayload =
-                    client.key_domain(proposal_payload(*proposal_id)).await?;
-                let proposal = Proposal {
-                    description,
-                    payload,
-                };
-                json(&proposal)?;
-            }
-            State { proposal_id } => {
-                let state: proposal::State =
-                    client.key_domain(proposal_state(*proposal_id)).await?;
-                json(&state)?;
-            }
-            Period { proposal_id } => {
-                let start: u64 = client
-                    .key_proto(proposal_voting_start(*proposal_id))
-                    .await?;
-                let end: u64 = client.key_proto(proposal_voting_end(*proposal_id)).await?;
-                let period = json!({
-                    "voting_start_block": start,
-                    "voting_end_block": end,
-                });
-                json(&period)?;
-            }
-            ValidatorVotes { proposal_id } => {
-                let voting_validators: validator::List =
-                    client.key_domain(voting_validators(*proposal_id)).await?;
-
-                let mut votes: BTreeMap<IdentityKey, Vote> = BTreeMap::new();
-                for identity_key in voting_validators.0.iter() {
-                    let vote: Vote = client
-                        .key_domain(validator_vote(*proposal_id, *identity_key))
-                        .await?;
-                    votes.insert(*identity_key, vote);
+            ProposalCmd::Proposal { proposal_id, query } => match query {
+                Definition => {
+                    let title: String = client.key_proto(proposal_title(*proposal_id)).await?;
+                    let description: String =
+                        client.key_proto(proposal_description(*proposal_id)).await?;
+                    let payload: ProposalPayload =
+                        client.key_domain(proposal_payload(*proposal_id)).await?;
+                    let proposal = Proposal {
+                        title,
+                        description,
+                        payload,
+                    };
+                    json(&proposal)?;
                 }
-                json(&votes)?;
-            }
-            Tally { proposal_id: _ } => {
-                todo!("vote tallying not yet implemented");
-            }
+                State => {
+                    let state: proposal::State =
+                        client.key_domain(proposal_state(*proposal_id)).await?;
+                    json(&state)?;
+                }
+                Period => {
+                    let start: u64 = client
+                        .key_proto(proposal_voting_start(*proposal_id))
+                        .await?;
+                    let end: u64 = client.key_proto(proposal_voting_end(*proposal_id)).await?;
+                    let period = json!({
+                        "voting_start_block": start,
+                        "voting_end_block": end,
+                    });
+                    json(&period)?;
+                }
+                ValidatorVotes => {
+                    let voting_validators: validator::List =
+                        client.key_domain(voting_validators(*proposal_id)).await?;
+
+                    let mut votes: BTreeMap<IdentityKey, Vote> = BTreeMap::new();
+                    for identity_key in voting_validators.0.iter() {
+                        let vote: Vote = client
+                            .key_domain(validator_vote(*proposal_id, *identity_key))
+                            .await?;
+                        votes.insert(*identity_key, vote);
+                    }
+                    json(&votes)?;
+                }
+                Tally => {
+                    todo!("vote tallying not yet implemented");
+                }
+            },
         }
 
         Ok(())
