+++ wallet/src/plan.rs
@@ -128,7 +128,7 @@ where
     let source_index: Option<AddressIndex> = source_address.map(Into::into);
     let notes_to_spend = view
         .notes(NotesRequest {
-            fvk_hash: Some(fvk.hash().into()),
+            account_id: Some(fvk.hash().into()),
             asset_id: Some((*STAKING_TOKEN_ASSET_ID).into()),
             address_index: source_index.map(Into::into),
             amount_to_spend: spend_amount,
